Añadida gestión completa de citas y creado solicitudes (sin gestión)

+++ Booking/Booking-war/src/java/com/booking/controllers/EditClassController.java
@@ -89,7 +89,7 @@ public class EditClassController implements Serializable {
 
         classId = FacesUtil.getParameter("class");
         if (classId != null) {
-            currentClass = classFacade.findClassOfOrganisation(Integer.valueOf(classId), organisation);
+            currentClass = classFacade.findClassOfOrganisation(Long.valueOf(classId), organisation);
 
             if (currentClass != null) {
                 selectedServiceId = currentClass.getService().getId();
